feat(backend): integrate Soroban SDK and resolve API gateway configuration
- Update KeyManager to generate both Nimbus OctetKeyPair (offline JWT) and Stellar KeyPair (Web3)
- Refactor StellarAdapter to implement strict Soroban transaction simulation lifecycle
- Align NimbusJwtSigner claims (riderId, routeId, iss) with offline mobile scanner requirements
- Add flyway-mysql dependency to resolve Spring Boot 3.2 / Flyway 9 database initialization failure
- Sync Java 21 compilation targets across parent and child Maven modules

+++ afiapass-sdk/src/main/java/org/afiapass/core/domain/data/models/Permit.java
@@ -1,5 +1,6 @@
 package org.afiapass.core.domain.data.models;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 import java.util.UUID;
 
@@ -10,7 +11,7 @@ public record Permit(
         UUID id,
         String riderId,
         String routeId,
-        double amount,
+        BigDecimal amount,
         Instant issuedAt,
         Instant expiresAt,
         String stellarTxHash
